requirements.txt
ag-ui-protocol==0.1.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alabaster==1.0.0
annotated-types==0.7.0
anthropic==0.67.0
anyio==4.10.0
argcomplete==3.6.2
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
banks==2.2.0
beautifulsoup4==4.13.5
boto3==1.40.32
botocore==1.40.32
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cohere==5.18.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
deepdiff==8.6.1
defusedxml==0.7.1
deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docutils==0.21.2
duckdb==1.3.2
duckdb-engine==0.17.0
et-xmlfile==2.0.0
eval-type-backport==0.2.2
executing==2.2.1
faker==37.8.0
fasta2a==0.3.5
fastavro==1.12.0
filelock==3.19.1
filetype==1.2.0
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
genai-prices==0.0.27
google-auth==2.40.3
google-genai==1.38.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
griffe==1.14.0
groq==0.31.1
grpcio==1.75.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.35.0
hyperframe==6.1.0
idna==3.10
imagesize==1.4.1
importlib-metadata==8.7.0
invoke==2.2.0
ipykernel==6.30.1
ipython==9.5.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.9
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.2
llama-index-cli==0.5.1
llama-index-core==0.14.2
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.1
llama-index-llms-ollama==0.7.3
llama-index-llms-openai==0.5.6
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.2.0
llama-parse==0.6.54
logfire==4.7.0
logfire-api==4.7.0
markdown-it-py==4.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mcp==1.14.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mem0ai==0.1.117
mistralai==1.9.10
mpmath==1.3.0
multidict==6.6.4
mygeotab==0.9.3
mypy-extensions==1.1.0
myst-parser==4.0.1
nest-asyncio==1.6.0
networkx==3.5
nexus-rpc==1.1.0
nltk==3.9.1
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
ollama==0.5.4
openai==1.107.3
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-httpx==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
orderly-set==5.5.0
packaging==25.0
pandas==2.2.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
portalocker==3.2.0
posthog==6.7.4
prompt-toolkit==3.0.52
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-ai==0.3.5
pydantic-ai-slim==0.3.5
pydantic-core==2.33.2
pydantic-evals==0.3.5
pydantic-graph==0.3.5
pydantic-settings==2.10.1
pydot==4.0.1
pygments==2.19.2
pyparsing==3.2.4
pypdf==6.0.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-rapidjson==1.21
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.1.0
qdrant-client==1.15.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
roman-numerals-py==3.1.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.8
sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlalchemy==2.0.41
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
striprtf==0.0.26
sympy==1.14.0
temporalio==1.17.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
triton==3.4.0
types-protobuf==6.30.2.20250914
types-python-dateutil==2.9.0.20250822
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0